requirements.txt
beautifulsoup4==4.12.3
langchain_community==0.3.14
openai==1.59.6
PyPDF2==3.0.1
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.37.1
